ultra-detailed documentation comments for all of the shared class library. Move nested classes to their own files
