+++ ValheimBackupShared/BO/BackupFrequency.cs
@@ -4,6 +4,17 @@ using System.ComponentModel;
 namespace ValheimBackup.BO
 {
     /// <summary>
+    /// Represents a time period for backups
+    /// 
+    /// Implements INotifyPropertyChanged interface and Newtonsoft.JSON serialization attributes
+    /// </summary>
+    public enum BackupPeriod : int
+    {
+        Minutes, Hours, Days, Weeks
+    }
+
+
+    /// <summary>
     /// Represents the frequency of occurance for a backup event in the form of "every {amount} {unit}".
     /// eg. every 2 days, every 30 minutes, every 1 week
     /// </summary>
@@ -13,6 +24,9 @@ namespace ValheimBackup.BO
         private int _amount;
         private BackupPeriod _period;
 
+        /// <summary>
+        /// Frequency amount (number of units)
+        /// </summary>
         [JsonProperty]
         public int Amount
         {
@@ -27,6 +41,9 @@ namespace ValheimBackup.BO
             }
         }
 
+        /// <summary>
+        /// Frequency period (units)
+        /// </summary>
         [JsonProperty]
         public BackupPeriod Period
         {
@@ -41,11 +58,23 @@ namespace ValheimBackup.BO
             }
         }
 
+        /// <summary>
+        /// Empty default constructor for JSON deserialization
+        /// </summary>
+        public BackupFrequency() { }
+
+        /// <summary>
+        /// Create new BackupFrequency from params
+        /// </summary>
+        /// <param name="amount">frequency amount</param>
+        /// <param name="period">frequency period</param>
         public BackupFrequency(int amount, BackupPeriod period)
         {
             this.Amount = amount;
             this.Period = period;
         }
+
+        #region INotifyPropertychanged
 
         public event PropertyChangedEventHandler PropertyChanged;
 
@@ -56,5 +85,7 @@ namespace ValheimBackup.BO
                 PropertyChanged(this, new PropertyChangedEventArgs(propertyName));
             }
         }
+
+        #endregion
     }
 }
