+++ ValheimBackupShared/BO/BackupSettings.cs
@@ -8,22 +8,6 @@ using ValheimBackupShared.Properties;
 namespace ValheimBackup.BO
 {
     /// <summary>
-    /// Represents a time period for backups
-    /// </summary>
-    public enum BackupPeriod : int
-    {
-        Minutes, Hours, Days, Weeks
-    }
-
-    /// <summary>
-    /// Represents a time period for cleanups
-    /// </summary>
-    public enum CleanupPeriod : int
-    {
-        Minutes, Hours, Days, Weeks, Copies
-    }
-
-    /// <summary>
     /// All or only specific worlds
     /// </summary>
     public enum WorldSelection : int
@@ -32,30 +16,20 @@ namespace ValheimBackup.BO
     }
 
     /// <summary>
-    /// Represents all the information about how and when to backup worlds for a specific server.
+    /// Represents all the information about how, when, where, and what worlds
+    /// to backup for a specific server.
     /// </summary>
     [JsonObject(MemberSerialization.OptIn)]
     public class BackupSettings : INotifyPropertyChanged
     {
-        private DateTime _lastBackup;
         private string _worldDirectory;
         private string _backupDirectory;
         private WorldSelection _worldSelection;
         private List<string> _selectedWorlds;
 
-        public DateTime LastBackup
-        {
-            get => _lastBackup;
-            set
-            {
-                if(_lastBackup != value)
-                {
-                    _lastBackup = value;
-                    NotifyPropertyChanged("LastBackup");
-                }
-            }
-        }
-
+        /// <summary>
+        /// The remote directory on this server containing the world files
+        /// </summary>
         [JsonProperty]
         public string WorldDirectory
         {
@@ -70,6 +44,9 @@ namespace ValheimBackup.BO
             }
         }
 
+        /// <summary>
+        /// The local directory to save backup files in
+        /// </summary>
         [JsonProperty]
         public string BackupDirectory
         {
@@ -84,6 +61,9 @@ namespace ValheimBackup.BO
             }
         }
 
+        /// <summary>
+        /// Determines whether to backup all worlds or specific worlds only
+        /// </summary>
         [JsonProperty]
         public WorldSelection WorldSelection
         {
@@ -98,6 +78,11 @@ namespace ValheimBackup.BO
             }
         }
 
+        /// <summary>
+        /// Contains a list of the worlds that should be backed up.
+        /// This property will be ignored unless the user has specifically
+        /// selected to only backup specific worlds.
+        /// </summary>
         [JsonProperty]
         public List<string> SelectedWorlds
         {
@@ -112,12 +97,27 @@ namespace ValheimBackup.BO
             }
         }
 
+        /// <summary>
+        /// The schedule for performing remote file backups
+        /// </summary>
         [JsonProperty]
         public BackupSchedule Schedule { get; set; }
 
+        /// <summary>
+        /// The schedule for performing backup file cleanup
+        /// </summary>
         [JsonProperty]
         public CleanupFrequency CleanupSchedule { get; set; }
 
+        /// <summary>
+        /// Returns a new instance of Backup settings with default properties.
+        /// Defaults:
+        /// Backup schedule: every 30 minutes
+        /// World directory: /save/worlds
+        /// Backup directory: read from application settings.
+        /// World selection: All
+        /// Cleanup schedule: Every 10 copies
+        /// </summary>
         public static BackupSettings Default
         {
             get
@@ -134,10 +134,53 @@ namespace ValheimBackup.BO
             }
         }
 
-        public BackupSettings()
-        {
+        /// <summary>
+        /// A list of file extensions that are valid for world servers
+        /// </summary>
+        private static readonly string[] VALID_EXTENSIONS = new string[] { ".db", ".fwl" };
 
+        /// <summary>
+        /// Empty default constructor for JSON deserialization
+        /// </summary>
+        public BackupSettings()  { }
+
+        /// <summary>
+        /// Determines whether a file should be backed up, based on it's name
+        /// and extension.
+        /// </summary>
+        /// <param name="fileName"></param>
+        /// <param name="extension"></param>
+        /// <returns>True if the file should be backed up, false otherwise.</returns>
+        public bool ShouldBackup(string fileName, string extension)
+        {
+            if (WorldSelection == WorldSelection.Specific)
+            {
+                if (!SelectedWorlds.Contains(fileName))
+                {
+                    //specific worlds enabled, and not in selected worlds.
+                    return false;
+                }
+            }
+            if (!VALID_EXTENSIONS.Contains(extension))
+            {
+                // wrong extension
+                return false;
+            }
+
+            // no checks failed, return true
+            return true;
         }
+
+        /// <summary>
+        /// Overrides ToString to provide human readable representation of settings.
+        /// </summary>
+        /// <returns>"backing up {WorldSelection} worlds {Schedule}"</returns>
+        public override string ToString()
+        {
+            return "backing up " + WorldSelection.ToString() + " worlds " + Schedule.ToString();
+        }
+
+        #region INotifyPropertychanged
 
         public event PropertyChangedEventHandler PropertyChanged;
 
@@ -149,29 +192,6 @@ namespace ValheimBackup.BO
             }
         }
 
-        private static readonly string[] VALID_EXTENSIONS = new string[] { ".db", ".fwl" };
-        public bool ShouldBackup(string fileName, string extension)
-        {
-            if (WorldSelection == WorldSelection.Specific) // specific worlds enabled
-            {
-                if(!SelectedWorlds.Contains(fileName)) // not in list
-                {
-                    return false;
-                }
-            }
-            if(!VALID_EXTENSIONS.Contains(extension)) // wrong extension
-            {
-                return false;
-            }
-
-            return true; // true if not tests failed
-        } 
-
-        public override string ToString()
-        {
-            return "backing up " + WorldSelection.ToString() + " worlds " + Schedule.ToString();
-        }
-
-        //TODO: add naming convention for backup files
+        #endregion
     }
 }
