+++ ValheimBackupShared/BO/Server.cs
@@ -5,24 +5,25 @@ using System.ComponentModel;
 
 namespace ValheimBackup.BO
 {
+    /// <summary>
+    /// Represents a Server and all the info needed to perform backups for it.
+    /// </summary>
     [JsonObject(MemberSerialization.OptIn)]
     public class Server : INotifyPropertyChanged
     {
-        private long _id = -1;
+        private long _id;
         private string _name;
         private string _description;
 
+        /// <summary>
+        /// The servers ID, derived as the binary representation of the current
+        /// DateTime. Used to associate all of this servers Backup instances
+        /// with it. Should be a unique value among each server.
+        /// </summary>
         [JsonProperty]
         public long Id
         {
-            get
-            {
-                if (_id == -1)
-                {
-                    _id = DateTime.Now.ToBinary();
-                }
-                return _id;
-            }
+            get => _id;
             set
             {
                 if(_id != value)
@@ -33,6 +34,9 @@ namespace ValheimBackup.BO
             }
         }
 
+        /// <summary>
+        /// Name of the server
+        /// </summary>
         [JsonProperty]
         public string Name {
             get => _name;
@@ -46,6 +50,9 @@ namespace ValheimBackup.BO
             }
         }
 
+        /// <summary>
+        /// Description of the server
+        /// </summary>
         [JsonProperty]
         public string Description {
             get => _description;
@@ -59,15 +66,26 @@ namespace ValheimBackup.BO
             }
         }
 
+        /// <summary>
+        /// Connection details for the servers FTP connection
+        /// </summary>
         [JsonProperty]
         public FtpConnectionInfo ConnectionInfo { get; set; }
 
+        /// <summary>
+        /// Settings for server backup process
+        /// </summary>
         [JsonProperty]
         public BackupSettings BackupSettings { get; set; }
 
+        /// <summary>
+        /// TODO: Computed property that gets all the backup files associated with this server.
+        /// </summary>
         public List<Backup> Backups { get; set; }
 
-        //computed properties
+        /// <summary>
+        /// Computed property that concatenations server name and description.
+        /// </summary>
         public string NameAndDescription
         {
             get
@@ -76,13 +94,37 @@ namespace ValheimBackup.BO
             }
         }
 
-        public Server(string name, string desc, FtpConnectionInfo connectionInfo, BackupSettings backupSettings)
+        /// <summary>
+        /// Create a new Server from specified params.
+        /// </summary>
+        /// <param name="name">server name</param>
+        /// <param name="desc">server description</param>
+        /// <param name="connectionInfo">FTP connection info</param>
+        /// <param name="backupSettings">backup settings</param>
+        /// <param name="id">(optional) server id - defaults to: <code>DateTime.Now.ToBinary()</code></param>
+        public Server(string name, string desc, FtpConnectionInfo connectionInfo, BackupSettings backupSettings, long id = -1L)
         {
+            if (id == -1)
+            {
+                id = DateTime.Now.ToBinary();
+            }
+            _id = id;
             this.Name = name;
             this.Description = desc;
             this.ConnectionInfo = connectionInfo;
             this.BackupSettings = backupSettings;
         }
+
+        /// <summary>
+        /// Override ToString to return human readable server details.
+        /// </summary>
+        /// <returns>"{Name} - {ConnectionInfo}"</returns>
+        public override string ToString()
+        {
+            return Name +  " - " + ConnectionInfo;
+        }
+
+        #region INotifyPropertychanged
 
         public event PropertyChangedEventHandler PropertyChanged;
 
@@ -94,9 +136,6 @@ namespace ValheimBackup.BO
             }
         }
 
-        public override string ToString()
-        {
-            return Name +  " - " + ConnectionInfo;
-        }
+        #endregion
     }
 }
