+++ ValheimBackupShared/ValheimBackupShared.csproj
@@ -45,17 +45,24 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="BO\Backup.cs" />
+    <Compile Include="BO\BackupFilePair.cs" />
+    <Compile Include="BO\BackupFrequency.cs" />
+    <Compile Include="BO\BackupSchedule.cs" />
     <Compile Include="BO\BackupSettings.cs" />
+    <Compile Include="BO\CleanupFrequency.cs" />
     <Compile Include="BO\FtpConnectionInfo.cs" />
     <Compile Include="BO\Server.cs" />
-    <Compile Include="Class1.cs" />
     <Compile Include="Data\BackupBuilder.cs" />
     <Compile Include="Data\BackupDataCleaner.cs" />
     <Compile Include="Data\BackupDataManager.cs" />
     <Compile Include="Data\DataManager.cs" />
     <Compile Include="Data\ServerDataManager.cs" />
+    <Compile Include="FTP\FtpException.cs" />
     <Compile Include="FTP\FtpFileInfo.cs" />
+    <Compile Include="FTP\FtpFileStringFormatException.cs" />
     <Compile Include="FTP\FtpManager.cs" />
+    <Compile Include="FTP\FtpRequestException.cs" />
+    <Compile Include="FTP\FtpResponseException.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Settings.Designer.cs">
       <AutoGen>True</AutoGen>
